\name{RaMoNA-package}
\alias{RaMoNA-package}
\alias{RaMoNA}
\docType{package}
\title{
\packageTitle{RaMoNA}
}
\description{
\packageDescription{RaMoNA}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{RaMoNA}
\packageIndices{RaMoNA}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{RaMoNA}

Maintainer: \packageMaintainer{RaMoNA}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
