\name{RaMoNA-package}
\alias{RaMoNA-package}
\alias{RaMoNA}
\docType{package}
\title{
\packageTitle{RaMoNA}
}
\description{
\packageDescription{RaMoNA}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{RaMoNA}
\packageIndices{RaMoNA}

}
\author{
\packageAuthor{RaMoNA}

Maintainer: \packageMaintainer{RaMoNA}
}
\references{
http://fiehnlab.ucdavis.edu/
}
\keyword{ database }
